requirements.txt
absl-py==1.1.0
aniso8601==9.0.1
attrs==22.1.0
cachetools==5.2.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.5
cycler==0.11.0
easydict==1.9
flasgger==0.9.5
Flask==2.1.2
Flask-RESTful==0.3.9
fonttools==4.33.3
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
imgviz==1.5.0
importlib-metadata==4.11.4
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
kiwisolver==1.4.3
labelImg==1.8.6
labelme==5.0.1
lxml==4.9.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.2.2
mistune==2.0.4
natsort==8.1.0
numpy==1.21.6
oauthlib==3.2.0
onnx==1.12.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
Pillow==9.1.1
pkgutil_resolve_name==1.3.10
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5-tools==5.15.4.3.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
PyYAML==6.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
QtPy==2.1.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
shiboken2==5.15.2.1
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.0
